Removed stray fclose calls; set correct types in ran*?.c files (need to setup x-platform stuff); created cross platform file ops

+++ efmsource/ran2.c
@@ -1,6 +1,7 @@
 /* ran2.c */
 #include <stdio.h>
 #include <math.h>
+#include <stdint.h>
 
 
 #define   IM1 2147483563
@@ -24,12 +25,12 @@ extern float	idum;
 
 float Ran2(float *idum) /* idum is initialized with a a number <0 */
 {
- int           j;
- long          k;
- static  long  idum2=123456789;
- static  long  iy=0;
- static long   iv[NTAB];
- float         temp;
+ int            j;
+ long           k;
+ static int32_t idum2=123456789;
+ static int32_t iy=0;
+ static int32_t iv[NTAB];
+ float          temp;
 
  if(*idum<=0){
    if(-(*idum)<1)*idum=1;
